pcgroups: Fix zombie consumer
When creating the consumer, if it failed with an "exists" failure, then
succeeded in creating it after the delete, `instance.startConsuming()` was not called.

This PR updates the API to make it clearer when there's a failure and success to ensure that
we always do the right thing on succes.

Fixes #37